TP4/requirements.txt
# Requirements pour le TP Jour 4 - Optimisation des modeles IA
# Detection d'anomalies avec Hugging Face et Azure ML

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Hugging Face
transformers>=4.30.0
datasets>=2.12.0
huggingface-hub>=0.15.0
accelerate>=0.20.0

# Vision Models
timm>=0.9.0  # PyTorch Image Models

# ONNX et ONNX Runtime
onnx>=1.14.0
onnxruntime>=1.15.0
# Pour GPU (optionnel) : onnxruntime-gpu>=1.15.0

# Azure ML (pour Partie 3-5)
azure-ai-ml>=1.8.0
azure-identity>=1.13.0
azureml-core>=1.51.0

# Manipulation de donnees
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Metriques et evaluation
scikit-learn>=1.3.0

# Utilitaires
tqdm>=4.65.0
requests>=2.31.0

# Pour les tests de charge (Partie 5)
locust>=2.15.0  # Load testing
fastapi>=0.100.0  # API REST
uvicorn>=0.23.0  # ASGI server

